requirements.txt
pyserial==3.5
matplotlib==3.7.1
pandas==2.0.3
flask==2.3.3 